requirements.txt
aiofiles==24.1.0
tqdm==4.67.0
flask==3.0.3
numpy==1.26.4
pywebview==5.3.2
waitress==3.0.1
requests==2.32.3
psutil==6.1.0
edge-tts==6.1.18
flask_sqlalchemy==3.1.1
flask_apscheduler==1.13.1
openai==1.54.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
wordninja==2.0.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pydantic==2.9.2
moviepy==1.0.3
decorator==4.4.1
mutagen==1.47.0